Change unsigned long to size_t in return type

+++ cbits/fpstring.c
@@ -73,7 +73,7 @@ unsigned char fps_minimum(unsigned char *p, size_t len) {
 }
 
 /* count the number of occurences of a char in a string */
-unsigned long fps_count(unsigned char *p, size_t len, unsigned char w) {
+size_t fps_count(unsigned char *p, size_t len, unsigned char w) {
     unsigned long c;
     for (c = 0; len-- != 0; ++p)
         if (*p == w)
